net/netpolicy: enable and disable NET policy
This patch introduces functions to enable and disable NET policy.

For enabling, it collects device and CPU information and setup CPU/queue
mapping. Also, for some drivers like i40e driver, it will get better
performance if setting IRQ affinity. This patch provides an option
irq_affinity for driver to set IRQ affinity in NET policy enabling.

For disabling, it removes the IRQ affinity (if set) and mapping
information.

np_lock should protect the enable and disable state. It will be done
later in this series.

Signed-off-by: Kan Liang <kan.liang@intel.com>